Decouple sampling_time from rescaling_time.
The output of all LEs and LVs is done every sampling_time instead of
every rescaling_time. The benettin step is still done every
rescaling_time but the ginelli step is not.

Instead of multiplying the CLVs with R every rescaling_time, we
accumulate the instantaneous Rs in prodR and multiply it once every
sampling_time. This may lead to ill-conditioned R matrices: to check.

The F/BLVs are still computed every time step but they are not written
out at this frequency.

The resulting CLVs are sampled less frequently but still correspond to
the instantaneous CLV of that time step. However, the CLEs are averaged
over the past sampling_time period.