Write out prop and R every rescaling_time anyway.
This is necessary because otherwise the accumulation of too many Rs with
very small and large eigenvalues will lead to numerical underflows in
the prodR matrix. The same holds for the prop matrices needed to compute
the FLVs.

However having the BLVs (and all LE/LV in general) less frequently does
not affect their accuracy.

Since the writeout functions need to be aware of the filesize and hence
the timestep, we have created a separate function for prop and vectors
since these are now written out at different frequencies.